requirements-defi.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.13.1
redis>=5.0.1
celery>=5.3.4
prometheus-client>=0.19.0
structlog>=23.2.0